+++ gleam/src/typ.rs
@@ -1074,7 +1074,7 @@ pub fn infer_module(
                 };
 
                 // Assert that the inferred type matches the type of any recursive call
-                unify(&rec, &typ).map_err(|e| convert_unify_error(e, &meta))?;
+                unify(&rec, &typ, &mut env).map_err(|e| convert_unify_error(e, &meta))?;
 
                 // Insert the function into the environment
                 let typ = generalise(typ, level);
@@ -1382,7 +1382,8 @@ pub fn infer(expr: UntypedExpr, level: usize, env: &mut Env) -> Result<TypedExpr
                 unify_pattern(&clause.pattern, &subject_type, level, env)?;
 
                 let then = infer(clause.then, level, env)?;
-                unify(&return_type, then.typ()).map_err(|e| convert_unify_error(e, then.meta()))?;
+                unify(&return_type, then.typ(), env)
+                    .map_err(|e| convert_unify_error(e, then.meta()))?;
                 typed_clauses.push(Clause {
                     meta: clause.meta,
                     pattern: clause.pattern,
@@ -1404,7 +1405,7 @@ pub fn infer(expr: UntypedExpr, level: usize, env: &mut Env) -> Result<TypedExpr
         } => {
             let head = infer(*head, level, env)?;
             let tail = infer(*tail, level, env)?;
-            unify(tail.typ(), &list(head.typ().clone()))
+            unify(tail.typ(), &list(head.typ().clone()), env)
                 .map_err(|e| convert_unify_error(e, &meta))?;
             Ok(Expr::Cons {
                 meta,
@@ -1478,7 +1479,7 @@ pub fn infer(expr: UntypedExpr, level: usize, env: &mut Env) -> Result<TypedExpr
             let tail = infer(*tail, level, env)?;
 
             let value_type = env.new_unbound_var(level);
-            unify(&value_type, value.typ()).map_err(|e| convert_unify_error(e, &meta))?;
+            unify(&value_type, value.typ(), env).map_err(|e| convert_unify_error(e, &meta))?;
 
             let tail_row_type = env.new_unbound_var(level);
             unify(
@@ -1486,6 +1487,7 @@ pub fn infer(expr: UntypedExpr, level: usize, env: &mut Env) -> Result<TypedExpr
                     row: Box::new(tail_row_type.clone()),
                 },
                 tail.typ(),
+                env,
             )
             .map_err(|e| convert_unify_error(e, &meta))?;
 
@@ -1531,7 +1533,8 @@ pub fn infer(expr: UntypedExpr, level: usize, env: &mut Env) -> Result<TypedExpr
                     tail: Box::new(other_fields_typ),
                 }),
             };
-            unify(&dummy_map, map.typ())
+
+            unify(&dummy_map, map.typ(), env)
                 .map_err(|e| convert_unify_error(set_map_row_type(e), &meta))?;
 
             Ok(Expr::MapSelect {
@@ -1579,7 +1582,7 @@ fn infer_module_select(
             tail: Box::new(other_fields_typ),
         }),
     };
-    unify(&dummy_module, module_type).map_err(|e| convert_unify_error(e, meta))?;
+    unify(&dummy_module, module_type, env).map_err(|e| convert_unify_error(e, meta))?;
 
     Ok(selected_field_typ)
 }
@@ -1603,19 +1606,20 @@ fn unify_pattern(pattern: &Pattern, typ: &Type, level: usize, env: &mut Env) -> 
             Ok(())
         }
 
-        Pattern::Int { meta, .. } => unify(&int(), typ).map_err(|e| convert_unify_error(e, &meta)),
+        Pattern::Int { meta, .. } => {
+            unify(&int(), typ, env).map_err(|e| convert_unify_error(e, &meta))
+        }
 
         Pattern::Float { meta, .. } => {
-            unify(&float(), typ).map_err(|e| convert_unify_error(e, &meta))
+            unify(&float(), typ, env).map_err(|e| convert_unify_error(e, &meta))
         }
 
         Pattern::String { meta, .. } => {
-            unify(&string(), typ).map_err(|e| convert_unify_error(e, &meta))
+            unify(&string(), typ, env).map_err(|e| convert_unify_error(e, &meta))
         }
 
-        Pattern::Nil { meta, .. } => {
-            unify(&list(env.new_unbound_var(level)), typ).map_err(|e| convert_unify_error(e, &meta))
-        }
+        Pattern::Nil { meta, .. } => unify(&list(env.new_unbound_var(level)), typ, env)
+            .map_err(|e| convert_unify_error(e, &meta)),
 
         Pattern::Cons {
             meta, head, tail, ..
@@ -1647,7 +1651,7 @@ fn unify_pattern(pattern: &Pattern, typ: &Type, level: usize, env: &mut Env) -> 
                         for (pattern, typ) in pattern_args.iter().zip(&args) {
                             unify_pattern(pattern, typ, level, env)?;
                         }
-                        unify(&retrn, typ).map_err(|e| convert_unify_error(e, &meta))
+                        unify(&retrn, typ, env).map_err(|e| convert_unify_error(e, &meta))
                     } else {
                         // TODO: Incorrect number of args given to constructor
                         unimplemented!()
@@ -1656,7 +1660,7 @@ fn unify_pattern(pattern: &Pattern, typ: &Type, level: usize, env: &mut Env) -> 
 
                 c @ Type::App { .. } => {
                     if pattern_args.is_empty() {
-                        unify(&c, typ).map_err(|e| convert_unify_error(e, &meta))
+                        unify(&c, typ, env).map_err(|e| convert_unify_error(e, &meta))
                     } else {
                         // Error: singleton given args
                         unimplemented!()
@@ -1682,7 +1686,8 @@ fn unify_pattern(pattern: &Pattern, typ: &Type, level: usize, env: &mut Env) -> 
                 let elems = (0..(elems.len()))
                     .map(|_| env.new_unbound_var(level))
                     .collect();
-                unify(&Type::Tuple { elems }, &typ).map_err(|e| convert_unify_error(e, &meta))?;
+                unify(&Type::Tuple { elems }, &typ, env)
+                    .map_err(|e| convert_unify_error(e, &meta))?;
                 unify_pattern(pattern, &typ, level, env)
             }
 
@@ -1747,7 +1752,7 @@ fn infer_call(
         .zip(args)
         .map(|(typ, arg): (&mut Type, _)| {
             let arg = infer(arg, level, env)?;
-            unify(typ, arg.typ()).map_err(|e| convert_unify_error(e, &meta))?;
+            unify(typ, arg.typ(), env).map_err(|e| convert_unify_error(e, &meta))?;
             Ok(arg)
         })
         .collect::<Result<_, _>>()?;
@@ -1898,7 +1903,7 @@ fn instantiate(typ: Type, ctx_level: usize, env: &mut Env) -> Type {
     go(typ, ctx_level, &mut hashmap![], env)
 }
 
-#[derive(Debug)]
+#[derive(Debug, PartialEq)]
 enum UnifyError {
     CouldNotUnify {
         expected: Type,
@@ -1918,7 +1923,7 @@ enum UnifyError {
     },
 }
 
-fn unify(t1: &Type, t2: &Type) -> Result<(), UnifyError> {
+fn unify(t1: &Type, t2: &Type, env: &mut Env) -> Result<(), UnifyError> {
     if t1 == t2 {
         return Ok(());
     }
@@ -1926,7 +1931,7 @@ fn unify(t1: &Type, t2: &Type) -> Result<(), UnifyError> {
     // Collapse right hand side type links. Left hand side will be collapsed in the next block.
     if let Type::Var { typ } = t2 {
         if let TypeVar::Link { typ } = &*typ.borrow() {
-            return unify(t1, typ);
+            return unify(t1, typ, env);
         }
     }
 
@@ -1954,7 +1959,7 @@ fn unify(t1: &Type, t2: &Type) -> Result<(), UnifyError> {
                 Ok(())
             }
 
-            Action::Unify(t) => unify(&t, t2),
+            Action::Unify(t) => unify(&t, t2, env),
 
             Action::Error => Err(UnifyError::CouldNotUnify {
                 expected: (*t1).clone(),
@@ -1964,7 +1969,7 @@ fn unify(t1: &Type, t2: &Type) -> Result<(), UnifyError> {
     }
 
     if let Type::Var { .. } = t2 {
-        return unify(t2, t1).map_err(flip_unify_error);
+        return unify(t2, t1, env).map_err(flip_unify_error);
     }
 
     match (t1, t2) {
@@ -1984,7 +1989,7 @@ fn unify(t1: &Type, t2: &Type) -> Result<(), UnifyError> {
         ) => {
             if m1 == m2 && n1 == n2 && args1.len() == args2.len() {
                 for (a, b) in args1.iter().zip(args2) {
-                    unify(a, b)?;
+                    unify(a, b, env)?;
                 }
                 Ok(())
             } else {
@@ -1998,7 +2003,7 @@ fn unify(t1: &Type, t2: &Type) -> Result<(), UnifyError> {
         (Type::Tuple { elems: elems1, .. }, Type::Tuple { elems: elems2, .. }) => {
             if elems1.len() == elems2.len() {
                 for (a, b) in elems1.iter().zip(elems2) {
-                    unify(a, b)?;
+                    unify(a, b, env)?;
                 }
                 Ok(())
             } else {
@@ -2023,9 +2028,9 @@ fn unify(t1: &Type, t2: &Type) -> Result<(), UnifyError> {
         ) => {
             if args1.len() == args2.len() {
                 for (a, b) in args1.iter().zip(args2) {
-                    unify(a, b)?;
+                    unify(a, b, env)?;
                 }
-                unify(retrn1, retrn2)
+                unify(retrn1, retrn2, env)
             } else {
                 Err(UnifyError::CouldNotUnify {
                     expected: (*t1).clone(),
@@ -2035,10 +2040,10 @@ fn unify(t1: &Type, t2: &Type) -> Result<(), UnifyError> {
         }
 
         (Type::Map { row: row1 }, Type::Map { row: row2 }) => {
-            unify(row1, row2).map_err(set_map_row_type)
+            unify(row1, row2, env).map_err(set_map_row_type)
         }
 
-        (Type::Module { row: row1 }, Type::Module { row: row2 }) => unify(row1, row2),
+        (Type::Module { row: row1 }, Type::Module { row: row2 }) => unify(row1, row2, env),
 
         (Type::RowNil, Type::RowNil) => Ok(()),
 
@@ -2073,15 +2078,15 @@ fn unify(t1: &Type, t2: &Type) -> Result<(), UnifyError> {
                 head: (*head2).clone(),
                 tail: (*tail2).clone(),
             };
-            let tail2 = rewrite_row(t2, label1.clone(), *head1.clone())?;
+            let tail2 = unify_row_field(t2, label1.clone(), *head1.clone(), env)?;
 
             if let Some(typ) = unbound {
                 if let TypeVar::Link { .. } = &*typ.borrow() {
-                    // TODO: I don't remember what is supposed to happen here.
+                    // TODO: Recursive row type
                     unimplemented!()
                 }
             }
-            unify(tail1, &tail2)
+            unify(tail1, &tail2, env)
         }
 
         (_, _) => Err(UnifyError::CouldNotUnify {
@@ -2107,7 +2112,22 @@ fn set_map_row_type(e: UnifyError) -> UnifyError {
     }
 }
 
-fn rewrite_row(row: Type, label1: String, head1: Type) -> Result<Type, UnifyError> {
+/// Takes a type (expected to be a Row) and searches within it for a field
+/// with a particular label.
+/// If found the type of the label is unified with the expected type,
+/// possibly returning a unification error.
+///
+/// If the unification is successful the row is returned minus the field
+/// that has been unified, so that the caller may continue to check other
+/// fields in the row without having to pass over the field that has already
+/// been checked.
+///
+fn unify_row_field(
+    row: Type,
+    label1: String,
+    head1: Type,
+    env: &mut Env,
+) -> Result<Type, UnifyError> {
     match row {
         Type::RowNil => Err(UnifyError::FieldNotFound {
             label: label1,
@@ -2116,10 +2136,10 @@ fn rewrite_row(row: Type, label1: String, head1: Type) -> Result<Type, UnifyErro
 
         Type::RowCons { label, head, tail } => {
             if label == label1 {
-                unify(&head1, &head)?;
+                unify(&head1, &head, env)?;
                 Ok(*tail)
             } else {
-                let tail = rewrite_row(*tail, label1, head1)?;
+                let tail = unify_row_field(*tail, label1, head1, env)?;
                 Ok(Type::RowCons {
                     label,
                     head,
@@ -2128,16 +2148,81 @@ fn rewrite_row(row: Type, label1: String, head1: Type) -> Result<Type, UnifyErro
             }
         }
 
-        Type::Var { typ } => match &*typ.borrow() {
-            TypeVar::Unbound { .. } => unimplemented!(),
+        Type::Var { typ } => {
+            let (new_typ, tail) = match &*typ.borrow() {
+                TypeVar::Unbound { level, .. } => {
+                    let tail = env.new_unbound_var(*level);
+                    let t = Type::RowCons {
+                        label: label1,
+                        head: Box::new(head1),
+                        tail: Box::new(tail.clone()), // return this var as it is the tail
+                    };
+                    (t, tail)
+                }
 
-            TypeVar::Link { typ } => rewrite_row(*typ.clone(), label1, head1),
+                TypeVar::Link { typ } => return unify_row_field(*typ.clone(), label1, head1, env),
 
-            _ => unimplemented!(), // Expected a row
-        },
+                _ => unimplemented!(), // Expected a row
+            };
+            *typ.borrow_mut() = TypeVar::Link {
+                typ: Box::new(new_typ),
+            };
 
-        _ => unimplemented!(), // Expected a row
+            Ok(tail)
+        }
+
+        _ => unimplemented!(), // TODO: Expected a row
     }
+}
+
+#[test]
+fn rewrite_row_test() {
+    let mods = std::collections::HashMap::new();
+    let mut env = Env::new(&mods);
+
+    let row = Type::RowCons {
+        label: "a".to_string(),
+        head: Box::new(int()),
+        tail: Box::new(Type::RowNil {}),
+    };
+    assert_eq!(
+        Ok(Type::RowNil {}),
+        unify_row_field(row, "a".to_string(), int(), &mut env)
+    );
+
+    let row = Type::RowCons {
+        label: "b".to_string(),
+        head: Box::new(int()),
+        tail: Box::new(Type::RowCons {
+            label: "a".to_string(),
+            head: Box::new(int()),
+            tail: Box::new(Type::RowNil {}),
+        }),
+    };
+    assert_eq!(
+        Ok(Type::RowCons {
+            label: "b".to_string(),
+            head: Box::new(int()),
+            tail: Box::new(Type::RowNil {}),
+        }),
+        unify_row_field(row, "a".to_string(), int(), &mut env)
+    );
+
+    let row = Type::RowCons {
+        label: "b".to_string(),
+        head: Box::new(int()),
+        tail: Box::new(env.new_unbound_var(1)),
+    };
+    assert_eq!(
+        Ok(Type::RowCons {
+            label: "b".to_string(),
+            head: Box::new(int()),
+            tail: Box::new(Type::Var {
+                typ: Rc::new(RefCell::new(TypeVar::Unbound { id: 13, level: 1 }))
+            }),
+        }),
+        unify_row_field(row, "a".to_string(), int(), &mut env)
+    );
 }
 
 fn flip_unify_error(e: UnifyError) -> UnifyError {
@@ -2159,7 +2244,6 @@ fn flip_unify_error(e: UnifyError) -> UnifyError {
 /// the type, thus ensuring the type will be correctly generalized.
 ///
 fn update_levels(typ: &Type, own_level: usize, own_id: usize) -> Result<(), UnifyError> {
-    // TODO: move this into the match block
     if let Type::Var { typ } = &typ {
         let new_value = match &*typ.borrow() {
             TypeVar::Link { typ, .. } => return update_levels(typ, own_level, own_id),
@@ -3390,6 +3474,14 @@ pub fn two() { one() + zero() }",
                     }
                   }",
             typ: "module { fn length(List(a)) -> Int }",
+        },
+        Case {
+            src: "external fn the_map({a = Int, b = Int}) -> Int = \"\" \"\"
+            pub fn go(map) {
+                    let a = map.a
+                    the_map(map)
+                  }",
+            typ: "module { fn go({ a = Int, b = Int }) -> Int }",
         },
         // % TODO: mutual recursion
         //    // % {