+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 - Error messages now display the path of the file containing the problem.
 - Maps and modules with erroneous extra fields now have a custom error
   message.
+- Rows with tails that are unbound type variables are now correctly unified in
+  the type system. This fixes a bug in which maps and modules may sometimes
+  fail to type check when there is no error.
 
 ## v0.1.0 - 2019-04-15